Move PhaseMultiSelectListFilter to projects/filters.py
Addresses PR review: extract the filter (and DEFAULT_ACTIVE_PROJECT_PHASES)
from admin.py into a dedicated projects/filters.py module. admin.py and the
tests now import from there; no behavior change.
